requirements.txt


Django==2.0
django-autoslug==1.9.3
django-crispy-forms==1.7.0
django-extensions==2.0.0
django-filter==1.1.0
django-js-asset==1.0.0
django-model-utils==3.1.1
django-mptt==0.9.0
django-mptt-admin==0.5.0
django-mptt-urls==2.0.3
django-widget-tweaks==1.4.1
djangorestframework==3.7.7
djangorestframework-recursive==0.1.2
pytz==2018.3
six==1.11.0
typing==3.6.4
